Raviart-Thomas Basis Functions

Raviart-Thomas basis functions are also known as Rao-Wilton-Glisson functions [RWG82] and commonly employed for ...

First Order

BEAST.raviartthomas — Function

raviartthomas(mesh, cellpairs::Array{Int,2})

Constructs the RT basis on the input mesh. The i-th RT basis function will represent a current distribution flowing from cell cellpairs[1,i] to cellpairs[2,i] on the mesh.

Returns an object of type RTBasis, which comprises both the mesh and pairs of Shape objects which corresponds to the cell pairs, containing the necsessary coefficients and indices to compute the exact basis functions when required by the solver.

raviartthomas(mesh)

Conducts pre-processing on the input mesh by extracting the cell edges, cell pairs and indices required to construct the RT basis on the mesh.

Calls raviartthomas(mesh::Mesh, cellpairs::Array{Int,2}), which constructs the RT basis on the mesh, using the cell pairs identified.

Returns the RT basis object.

Triangles

Details on definition.

Quadrilaterals

Details on definition.

Higher Order

... same as GWP?