Plane Wave Excitation

A plane wave can be used as excitation in time-harmonic and time-domain scenarios.

Time-Harmonic

A time-harmonic plane wave with amplitude $a$, wave vector $\bm k = k \hat{\bm k}$, and polarization $\hat{\bm p}$ (vectors with a hat denote unit vectors) is defined by the field

\[\bm e_\mathrm{PW}(\bm x) = a \hat{\bm p} \, \mathrm{e}^{-\mathrm{j} \bm k \cdot \bm x} \,,\]

where the polarization and wave vector are orthogonal, that is,

\[\bm k \cdot \hat{\bm p} = 0\]

holds.

API

Warning

There are currently two APIs for the BEM plane wave. Fix in future.

BEAST.planewavemw3d — Function

planewavemw3d(;direction, polarization, wavenumber, gamma[, amplitude=1])

Create a plane wave solution to Maxwell's equations.

BEAST.Maxwell3D.planewave — Function

planewave(;
        direction    = error("missing arguement `direction`"),
        polarization = error("missing arguement `polarization`"),
        wavenumber   = error("missing arguement `wavenumber`"),
        amplitude    = one(real(typeof(wavenumber))))

Time-harmonic plane wave.

BEAST.planewavevie — Function

planewavevie(; direction = error("missing arguement direction"), polarization = error("missing arguement polarization"), wavenumber = error("missing arguement wavenumber"), amplitude = 1, )

For volume integral equations

Time-Domain

A plane wave in the time-domain is defined as ...

API

Some more details would be helpful here.

BEAST.planewave — Function

planewave(polarisation,direction,amplitude,speedoflight)

Time-domain plane wave.